feat: compressImage
